[Adjustment][Maxell] moved configs
